+++ Controls/buttons.ts
@@ -6,10 +6,10 @@
 
 import * as ButtonTemplate from 'wml!Controls/_buttons/ButtonBase';
 
-export {default as Button, simpleCssStyleGeneration, IViewMode, defaultHeight, defaultFontColorStyle, getDefaultOptions} from './_buttons/Button';
+export {default as Button, simpleCssStyleGeneration, defaultHeight, defaultFontColorStyle, getDefaultOptions} from './_buttons/Button';
 export {default as ArrowButton, IArrowButtonOptions} from './_buttons/ArrowButton';
 export {default as ActualApi} from './_buttons/ActualApi';
 export {IClick as IClick} from './_buttons/interface/IClick';
-export {IButton as IButton} from './_buttons/interface/IButton';
+export {IButton as IButton, IViewMode} from './_buttons/interface/IButton';
 export {IButtonOptions as IButtonOptions} from './_buttons/interface/IButton';
 export {ButtonTemplate};
